large changes
-altered environment tint
-removed day/night cycle
-added combat vfx numbers
-added support for boss/healing/shop rooms
-finished money system

+++ Assets/Scripts/EnemyWalkerAI.cs
@@ -88,7 +88,7 @@ public class EnemyWalkerAI : MonoBehaviour
         }
     }
 
-    public void TakeDamage (float dmg) 
+    public void TakeDamage (int dmg) 
     {
         health -= dmg;
         //StartCoroutine(Stun(0.25f, 12f));
